create login as common service
