+++ Zebra.AuthService.API/Controllers/ClientController.cs
@@ -6,7 +6,6 @@ using System.Security.Claims;
 using System.Threading.Tasks;
 using Zebra.AuthService.API.Models;
 using Zebra.AuthService.API.Resources;
-using Zebra.AuthService.API.Services.Token;
 
 namespace Zebra.AuthService.API.Controllers
 {
@@ -15,19 +14,13 @@ namespace Zebra.AuthService.API.Controllers
     public class ClientController : ControllerBase
     {
         private readonly UserManager<IdentityUser> _userManager;
-        private readonly SignInManager<IdentityUser> _signInManager;
-        private readonly ICreateToken _createToken;
         private readonly IStringLocalizer<ClientLocalizer> _localizer;
 
         public ClientController(
             UserManager<IdentityUser> userManager, 
-            SignInManager<IdentityUser> signInManager,
-            ICreateToken createToken,
             IStringLocalizer<ClientLocalizer> localizer)
         {
             _userManager = userManager;
-            _signInManager = signInManager;
-            _createToken = createToken;
             _localizer = localizer;
         }
 
@@ -70,31 +63,6 @@ namespace Zebra.AuthService.API.Controllers
             await _userManager.AddClaimAsync(registeredUser, claim);
 
             return Ok();
-        }
-
-        [HttpPost]
-        [Route("login")]
-        public async Task<IActionResult> Login([FromBody] RegisterLoginClientApiModel user)
-        {
-            var identityUser = await _userManager.FindByEmailAsync(user.Email);
-
-            if (identityUser is null)
-            {
-                return BadRequest(_localizer["IncorrectPasswordOrEmail"].Value);
-            }
-
-            var loginResult = await _signInManager.CheckPasswordSignInAsync(identityUser, user.Password, false);
-
-            if (!loginResult.Succeeded)
-            {
-                return BadRequest(_localizer["IncorrectPasswordOrEmail"].Value);
-            }
-
-            var claims = await _userManager.GetClaimsAsync(identityUser);
-
-            var token = _createToken.Create(claims);
-
-            return Ok(token);
         }
     }
 }
